migrations/app/19_20240708142702_update.py
# -*- coding: utf-8 -*-
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "mrg__patientaddress" ADD "created_at" TIMESTAMPTZ NOT NULL  DEFAULT CURRENT_TIMESTAMP;
        ALTER TABLE "mrg__patientcns" ADD "created_at" TIMESTAMPTZ NOT NULL  DEFAULT CURRENT_TIMESTAMP;
        ALTER TABLE "mrg__patienttelecom" ADD "created_at" TIMESTAMPTZ NOT NULL  DEFAULT CURRENT_TIMESTAMP;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "mrg__patientcns" DROP COLUMN "created_at";
        ALTER TABLE "mrg__patientaddress" DROP COLUMN "created_at";
        ALTER TABLE "mrg__patienttelecom" DROP COLUMN "created_at";"""
