pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "API for FHIR Utils"
authors = ["Escritório de Dados <escritoriodedados@gmail.com>"]
license = "MIT"
homepage = "https://github.com/prefeitura-rio/fhir-utils"
repository = "https://github.com/prefeitura-rio/fhir-utils"
keywords = ["fhir", "utilities", "data", "health", "government"]

[tool.poetry.dependencies]
python = "^3.11"
aerich = "^0.7.2"
aioredis = "^2.0.1"
bcrypt = "<4.1"
fastapi = "0.109.2"
loguru = "^0.7.0"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "0.0.7"
sentry-sdk = { version = "^1.37.1", extras = ["fastapi"] }
tortoise-orm = { extras = ["asyncpg"], version = "0.19.3" }
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
httpx = "^0.26.0"
pytest = "^7.4.4"
infisical = "1.5.0"
pandas = "^2.1.4"
validate-docbr = "^1.10.0"
pyjwt = "^2.8.0"
urllib3 = "2.0.7"
idna = "3.7"
basedosdados = "^2.0.0b16"


[tool.poetry.group.dev.dependencies]
black = {version = "^24.4.2", allow-prereleases = true}
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
pytest-asyncio = "^0.21.1"
taskipy = "^1.12.2"
isort = "^5.12.0"
pytest-cov = "^4.1.0"
pytest-ordering = "^0.6"

[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"

[tool.coverage.run]
omit = ["app/config/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["dist", "build", ".tox", ".eggs", "scripts"]
addopts = ["--doctest-modules", "--cov=app", "-ra", "-v"]

[tool.taskipy.tasks]
create_user = "python scripts/create_user.py"
lint = "black . && isort . && flake8 ."
serve = "uvicorn app.main:app --reload --port 8001"
test = "pytest"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
